Update patch set 6
Patch Set 6:

Regarding seeking, an issue I've had since the beginning of CM 12.1, is that when starting to play a movie in mkv format, and then for example jumping to the middle of the movie, there's a pause for several seconds before the playback resumes. This issue is not present when testing with awesomeplayer, and was not present with nuplayer in 12.0 either.

Patch-set: 6