Update patch set 6
Patch Set 6:

Arne, is this just with MKV?  It's probably unrelated to this code unless you've got some weird format. Do you have a test file? And also, what chipset is this on and what is the value of mm.enable.qcom_parser? They have an MKV extractor but it sucks really bad and the one we have in the framework can play more formats.

Also, the seek issue specific to FFMPEG is fixed in the latest patchset. Still have an issue with some formats giving a black thumbnail.

Patch-set: 6