Update patch set 6
Patch Set 6:

I tested with a movie in AVI format, and this format seems fine as the playback resumed instantly when jumping to the middle of it. So it's probably just MKV files. The particular MKV file which I have the issue in, is a 2 hour movie. Filename: World.War.Z.2013.UNRATED.BDRip.X264-SPARKS.mkv

It also seems that the further into the movie the jump is, the longer it takes to resume playback. I started playing from the beginning, and then jumped to the very end, and it took well over 10 seconds for it to resume.

Chipset is msm8930 and mm.enable.qcom_parser=3407871
But as I said, no issue in CM 12.0 with the same file.

Patch-set: 6