Update patch set 6
Patch Set 6:

Also, change your mm-parser property to 3310129. The current value you have in there is enabling EVERYTHING in mm-parser and it kind of sucks for a lot of things..

If you convert it to a binary value, it represents these flags:
#define PARSER_AAC    0x00000001
#define PARSER_AC3    0x00000002
#define PARSER_AMR_NB 0x00000004
#define PARSER_AMR_WB 0x00000008
#define PARSER_ASF    0x00000010
#define PARSER_AVI    0x00000020
#define PARSER_DTS    0x00000040
#define PARSER_FLV    0x00000080
#define PARSER_3GP    0x00000100
#define PARSER_3G2    0x00000200
#define PARSER_MKV    0x00000400
#define PARSER_MP2PS  0x00000800
#define PARSER_MP2TS  0x00001000
#define PARSER_MP3    0x00002000
#define PARSER_OGG    0x00004000
#define PARSER_QCP    0x00008000
#define PARSER_WAV    0x00010000

And the two highest bits are for DivX support. If you don't have WMV/WMA support on your device, also disable the ASF bits from the value I gave you.

Patch-set: 6