Update patch set 1
Patch Set 1:

Pretty sure the failure is in FFMPEG. It's getting a port setting change which is reading width/height from the AVCodecContext. This is resulting in the width/height being set to 1/2 of what it should be. Unsure why, but FWIW, both failing codecs (MPEG4 and H263) share the same decode routine (ff_h263_video_decode) in libavcodec.

Patch-set: 1