Update patch set 3
Patch Set 3:

This also adds significantly to our CTS run time. We now run a lot of the tests an additional time, once per codec that declares support for a particular codec. In addition to this being a waste of time, the FFMPEG codecs also fail the adaptive playback tests.

Patch-set: 3