Update patch set 1
Patch Set 1:

I'd investigate the fallback code in Stagefright since it's functionality that we added to allow this stuff to work. In AOSP, if the HW codec fails, it will just stop there. In CM, we will retry using the remaining decoders. Of course I am absolutely confident there are zero bugs in this code :)

Patch-set: 1