Update patch set 3
Patch Set 3:

The short answer is no .The oggextractor will always give a fixed confidence level 0.2f. And it is a very lightweight checking for a 4 bytes headers and may be wrong. Therefore it was giving a low confidence level. Actually it can be another media file type such as mp3/midi etc. Giving it a 0.2f by aosp allows others aosp extractor to use instead of oggextractor if it have a high confidence level > 0.2f.

Patch-set: 3