Update patch set 2
Patch Set 2:

No, this intentionally make start time reported by kKeyTime  to be always be zero when the video starts. The packet timestamp we retrieved (pktTS) should have stream  start time added to it already.
And we add back the start time during seek (seekPTS += startTimeUs;). since the seek time from frameworks/av is starts from 0.

This definite will break the seek at least for video will a start time not equal to zero. And the display of the current time will not be correct in UI, since frameworks/av assume the video start time always be 0 (kKeyTime =0 when video starts).

Patch-set: 2