Update patch set 2
Patch Set 2:

Do you know which video cause the problem ?
That patch does not seems to be correct for me.
The packet timestamp we retrieved (pktTS) should have stream  start time added to it already. So the intention of that is just to minus the start time from packet time in order to have kKeyTime always started with 0 no matter the stream start time is.

Patch-set: 2
Label: Code-Review=0