Update patch set 1
Patch Set 1:

Note: This makes a big difference in what extractors we use.  As far as I can see in the code, prior to this change we were using the ffmpeg extractor for all media files except for DRM content (DRMExtractor / WVMExtractor) and after this change we are only using it for any additional filetypes that ffmpeg ends up adding support for.

Rerunning some CTS tests to make sure that there are no regressions and I expect it to fix about 24 failing tests (per arch).

The main reason to do this though is to keep our media more consistent with aosp rather than further from aosp unless we specifically know something is better in ffmpeg.

Patch-set: 1
Label: Verified=0