Update patch set 2
Patch Set 2:

Based on pme, prior to this change I had 123 failures in CtsMediaTestCases and after this change I had 60.  There are new failures but my immediate skim suggests that they are random failures and some of the tests that now pass are randomly passing due to the change (and not really fixing the same underlying issue that caused it to fail with the old code and pass with the new code).

This is based on a run today on PS2 which fixed a NPE which was causing some failures that were definitely reproducible.

Patch-set: 2