Update patch set 3
Patch Set 3:

Dang it!  I tested that but (stupidly) did so using Chrome which I seem to have a hard time remembering doesn't actually use the normal media stack.  Sigh.

See commit:

 https://review.lineageos.org/175682

(and the one before it to properly support opus) and let me know if that fixes the problem for you.

Patch-set: 3