ffmpeg-extractor: Use our extractor for codecs not supported by SF
In commit:

FFmpegExtractor: Don't use our extractor when we agree with StageFright

we disable our extractor if we sniffed the same mime type as StageFright.
However it turns out that we broke support for anything that we want
to support via ffmpeg such as 96 KHz flac files.

This commit enforces the real intent of the previous commit which was to
use AOSP extractors for AOSP codecs and to use ffmpeg for ffmpeg
supported codecs.

Change-Id: I8e3be9d813719d570ca015ea0441735253f258bc