Update patch set 3
Patch Set 3:

Arne,

Did you try to commit I linked in the comment immediately before yours?  I can't tell from your partial logcat if it is the same problem or not (I'm not seeing the log messages for detecting the audio source, only the video source).

If this isn't fixed by my followup commit, can you provide an example that fails (but used to work) so I can reproduce?  I have tried some random samples from the internet and the hevc files in cts and I don't have any examples that are currently broken but previously worked but there are a lot of different media variations out there.

Patch-set: 3